requirements.txt
aiohttp>=3.6.2
backtrader>=1.9.76.123
boto3>=1.14.10
multiprocessing_logging>=0.3.1
pyarrow>=1.0.1
snowflake-connector-python
alpaca-backtrader-api
trading_calendars>=1.11.11
kafka-python>=2.0.1